switch to `src/`-layout
